requirements_prod.txt
aiofiles==23.2.1
aiosqlite==0.19.0
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mirakuru==2.5.1
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
port-for==0.7.1
psutil==5.9.5
psycopg==3.1.12
pycodestyle==2.11.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyflakes==3.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.21
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
tomli==1.2.3
typing_extensions==4.8.0
uvicorn==0.23.2
